requirements.txt
fastapi
uvicorn
pillow
imagehash
pandas
exifread
python-multipart
